re-instate important check lost in 016a075e79

+++ nouveau/lucene4/src/main/java/org/apache/couchdb/nouveau/lucene4/core/Lucene4Index.java
@@ -137,9 +137,11 @@ public class Lucene4Index extends Index<IndexableField> {
                     writer.rollback();
                 },
                 () -> {
-                    var dir = writer.getDirectory();
-                    for (final String name : dir.listAll()) {
-                        dir.deleteFile(name);
+                    if (isDeleteOnClose()) {
+                        var dir = writer.getDirectory();
+                        for (final String name : dir.listAll()) {
+                            dir.deleteFile(name);
+                        }
                     }
                 });
     }
